split the engine up into @opensouls/engine as a library and soul-engine as a CLI command
