--- packages/engine/bin/run.js
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-
-import { run } from '../dist/cli/index.mjs';
-
-run()
