+++ packages/soul-engine-cli/bin/dev.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env npx tsx
 /* eslint-disable node/shebang */
-import { run } from '../src/cli/index';
+import { run } from '../src/index';
 
 run()
