+++ packages/soul-engine-cli/bin/run.cmd
