+++ packages/soul-engine-cli/bin/run.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+import { run } from '../dist/index.mjs';
+
+run()
