+++ packages/soul-engine-cli/etc/engine.api.md
@@ -1,0 +1,235 @@
+## API Report File for "@opensouls/engine"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+import { DeveloperInteractionRequest } from '@opensouls/core';
+import { InternalPerception } from '@opensouls/core';
+import { Json } from '@opensouls/core';
+import { Perception } from '@opensouls/core';
+import { SoulEnvironment } from '@opensouls/core';
+import { WorkingMemory } from '@opensouls/core';
+
+// @public (undocumented)
+export const ALLOWED_RAG_FILE_EXTENSIONS: string[];
+
+// @public (undocumented)
+export interface Blueprint {
+    // (undocumented)
+    context: string;
+    // (undocumented)
+    defaultEnvironment?: SoulEnvironment;
+    // (undocumented)
+    entity: string;
+    // (undocumented)
+    initialProcess: MentalProcess<any>;
+    // (undocumented)
+    mentalProcesses: MentalProcess<any>[];
+    // (undocumented)
+    name: string;
+    // (undocumented)
+    subprocesses?: MentalProcess<any>[];
+}
+
+// @public (undocumented)
+export type CognitiveEvent = CognitiveEventAbsolute | CognitiveEventOffset;
+
+// @public (undocumented)
+export interface CognitiveEventAbsolute extends CognitiveEventBase {
+    // (undocumented)
+    when: Date;
+}
+
+// @public (undocumented)
+export interface CognitiveEventBase {
+    // (undocumented)
+    params?: Json;
+    // (undocumented)
+    perception: Omit<InternalPerception, "_id" | "_kind" | "_pending" | "_timestamp" | "internal">;
+    // (undocumented)
+    process: MentalProcess<any>;
+}
+
+// @public (undocumented)
+export interface CognitiveEventOffset extends CognitiveEventBase {
+    // (undocumented)
+    in: number;
+}
+
+// @public (undocumented)
+export interface DefaultActions {
+    // (undocumented)
+    dispatch: (evt: DeveloperInteractionRequest) => void;
+    // (undocumented)
+    expire: () => void;
+    // (undocumented)
+    log: (...args: any) => void;
+    // (undocumented)
+    scheduleEvent: (evt: CognitiveEvent) => void;
+    // (undocumented)
+    speak: (message: AsyncIterable<string> | string) => void;
+}
+
+// @public (undocumented)
+export const defaultRagBucketName: (blueprint: string) => string;
+
+// @public (undocumented)
+export type Embedding = number[];
+
+// @public (undocumented)
+export type MentalProcess<ParamType = Record<number | string, any>, CortexStepType = any> = (args: MentalProcessArguments<ParamType, CortexStepType>) => Promise<CortexStepType | WorkingMemory>;
+
+// @public (undocumented)
+export interface MentalProcessArguments<ParamType, CortexStepType = any> {
+    // (undocumented)
+    params: ParamType;
+    // (undocumented)
+    step: CortexStepType;
+    // (undocumented)
+    workingMemory: WorkingMemory;
+}
+
+// @public (undocumented)
+export interface RagConfigfile {
+    // (undocumented)
+    bucket: string;
+}
+
+// @public (undocumented)
+export interface RagIngestionBody {
+    // (undocumented)
+    content: string;
+    // (undocumented)
+    contentType?: string;
+    // (undocumented)
+    maxTokens?: number;
+    // (undocumented)
+    metadata?: VectorMetadata;
+    // (undocumented)
+    rootKey: string;
+}
+
+// @public (undocumented)
+export interface RagSearchOpts {
+    // (undocumented)
+    bucketName?: string;
+    // (undocumented)
+    limit?: number;
+    // (undocumented)
+    maxDistance?: number;
+    // (undocumented)
+    query: Embedding | string;
+}
+
+// @public (undocumented)
+export interface SoulConfig {
+    // (undocumented)
+    path?: string;
+    // (undocumented)
+    paths?: string[];
+    // (undocumented)
+    soul: string;
+}
+
+// @public
+export interface SoulHooks {
+    // (undocumented)
+    useActions: () => DefaultActions;
+    // (undocumented)
+    usePerceptions: () => {
+        invokingPerception: Perception | null | undefined;
+        pendingPerceptions: {
+            current: Perception[];
+        };
+    };
+    // (undocumented)
+    useProcessManager: () => {
+        invocationCount: number;
+        setNextProcess: <PropType>(process: MentalProcess<PropType>, props?: PropType) => void;
+        wait: (ms: number) => Promise<void>;
+    };
+    // (undocumented)
+    useProcessMemory: <T = null>(initialValue: T) => {
+        current: T;
+    };
+    // (undocumented)
+    useRag(bucketName?: string): {
+        search: (opts: RagSearchOpts) => Promise<VectorRecordWithSimilarity[]>;
+        withRagContext: <T = any>(step: T, opts?: WithRagContextOpts) => Promise<T>;
+    };
+    // (undocumented)
+    useSoulMemory: <T = null>(name: string, initialValue?: T) => {
+        current: T;
+    };
+    // (undocumented)
+    useSoulStore: () => {
+        createEmbedding: (content: string) => Promise<Embedding>;
+        delete: (key: string) => void;
+        get: <T = unknown>(key: string, opts?: SoulStoreGetOpts) => (typeof opts extends {
+            includeMetadata: true;
+        } ? VectorRecord : T) | undefined;
+        search: (query: Embedding | string, filter?: VectorMetadata) => Promise<VectorRecordWithSimilarity[]>;
+        set: (key: string, value: Json, metadata?: VectorMetadata) => void;
+    };
+}
+
+// @public (undocumented)
+export interface SoulStoreGetOpts {
+    // (undocumented)
+    includeMetadata?: boolean;
+}
+
+// @public (undocumented)
+export const useActions: SoulHooks["useActions"];
+
+// @public (undocumented)
+export const usePerceptions: SoulHooks["usePerceptions"];
+
+// @public (undocumented)
+export const useProcessManager: SoulHooks["useProcessManager"];
+
+// @public (undocumented)
+export const useProcessMemory: SoulHooks["useProcessMemory"];
+
+// @public (undocumented)
+export const useRag: (bucketName?: string) => any;
+
+// @public (undocumented)
+export const useSoulMemory: SoulHooks["useSoulMemory"];
+
+// @public (undocumented)
+export const useSoulStore: SoulHooks["useSoulStore"];
+
+// @public (undocumented)
+export type VectorMetadata = Record<string, Json>;
+
+// @public (undocumented)
+export interface VectorRecord {
+    // (undocumented)
+    content: Json;
+    // (undocumented)
+    embedding?: Embedding;
+    // (undocumented)
+    key: string;
+    // (undocumented)
+    metadata: VectorMetadata;
+}
+
+// @public (undocumented)
+export interface VectorRecordWithSimilarity extends VectorRecord {
+    // (undocumented)
+    similarity: number;
+}
+
+// @public (undocumented)
+export interface WithRagContextOpts {
+}
+
+
+export * from "@opensouls/core";
+export * from "@opensouls/soul";
+
+// (No @packageDocumentation comment for this package)
+
+```
