+++ packages/soul-engine-cli/etc/soul-engine.api.md
@@ -1,0 +1,15 @@
+## API Report File for "soul-engine"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+import { Command } from 'commander';
+
+// @public (undocumented)
+const run_2: () => Command;
+export { run_2 as run }
+
+// (No @packageDocumentation comment for this package)
+
+```
