+++ packages/soul-engine-cli/package.json
@@ -1,6 +1,9 @@
 {
-  "name": "@opensouls/engine",
-  "version": "0.0.20",
+  "name": "soul-engine",
+  "version": "0.1.0",
+  "bin": {
+    "soul-engine": "./bin/run.js"
+  },
   "type": "module",
   "description": "Write your own souls to run on the OPEN SOULS soul-engine",
   "main": "dist/index.mjs",
@@ -31,9 +34,14 @@
   },
   "dependencies": {
     "@hocuspocus/provider": "=2.9.0",
-    "@opensouls/core": "^0.0.20",
-    "@opensouls/soul": "^0.0.20",
-    "handlebars": "^4.7.8"
+    "@opensouls/engine": "^0.0.20",
+    "commander": "^12.0.0",
+    "conf": "^12.0.0",
+    "cross-fetch": "^4.0.0",
+    "execa": "^8.0.1",
+    "glob": "^10.3.10",
+    "handlebars": "^4.7.8",
+    "open": "^10.1.0"
   },
   "devDependencies": {
     "@microsoft/api-extractor": "^7.43.0",
