+++ packages/soul-engine-cli/scripts/build.ts
@@ -2,7 +2,6 @@ import { $ } from 'execa'
 import esbuild from 'esbuild'
 import { Extractor, ExtractorConfig, ExtractorResult } from '@microsoft/api-extractor';
 import { join } from 'path';
-import { readFile, appendFile } from "node:fs/promises"
 
 await $`rm -rf dist lib temp`
 await $`mkdir dist`
@@ -39,19 +38,13 @@ const defaultParams:esbuild.BuildOptions = {
   outExtension: { ".js": ".cjs" },
 }
 
+// build the CLI/commands
 await esbuild.build({
   ...defaultParams,
+  entryPoints: ['src/index.ts'],
+  outdir: 'dist',
   format: 'esm',
   outExtension: { ".js": ".mjs" },
 })
-
-await esbuild.build({
-  ...defaultParams,
-  format: 'cjs',
-})
-
-// see https://github.com/microsoft/rushstack/issues/1709
-const globals = await readFile("src/globals.d.ts", "utf-8")
-await appendFile("dist/types.d.ts", globals)
 
 await $`rm -rf lib temp`
