+++ packages/soul-engine-cli/src/commands/init.ts
