+++ packages/soul-engine-cli/src/commands/logout.ts
