+++ packages/soul-engine-cli/src/commands/rag/index.ts
