+++ packages/soul-engine-cli/src/commands/rag/watch.ts
