+++ packages/soul-engine-cli/src/debugChat/index.ts
