+++ packages/soul-engine-cli/src/fileSystem/file-watcher.ts
