+++ packages/soul-engine-cli/src/sockets/soul-engine-socket.ts
